+++ src/lib/git-off-my-land-lib.js
@@ -22,7 +22,7 @@ async function filterFilesList(rawStdOut: string, ignoreGitStatusResultPrefixes:
     let committedFilenames = new Set();
 
     try 
-    {
+    {        
         const committedFilesArray = rawStdOut.trim().split(EOLChar);
 
         committedFilesArray.forEach((f) => 
@@ -64,47 +64,89 @@ async function filterFilesList(rawStdOut: string, ignoreGitStatusResultPrefixes:
 
 async function scanFilteredFiles(committedFiles: Set, fileContentRegexps: Array, violatingFilenameExtensions: Array, filesToIgnore: Array)
 {
-    let err;
-    let violations = [];
+    let err;    
+    let filteredViolations = [];  
 
     try
     {
+        let rawViolations = [];
         const committedFilesArray = [...committedFiles];
 
         // Note: The method for handling async/await in a for loop is from https://blog.lavrton.com/javascript-loops-how-to-handle-async-await-6252dd3c795
         for(let i in committedFilesArray)
         {
             const committedFile = committedFilesArray[i];
-
             const stats = await stat(committedFile);
-
-            if(stats.isFile()) // Check we're not going to try to read a dir as that would bomb
+            if(stats.isFile()) // Check we're not going to try to read a dir because that would bomb
             {
-                const content = await readFile(committedFile, "utf8");  
+                if(rawViolations[committedFile] === undefined)
+                {
+                    rawViolations[committedFile] = 
+                    {
+                        content: [],
+                        extension: []
+                    };
+                }
 
+                // Content-based scanning
+                const content = await readFile(committedFile, "utf8");  
                 for(let j in fileContentRegexps)
                 {
                     const pattern = fileContentRegexps[j];
                             
                     if(content.match(pattern.regexp) && filesToIgnore.includes(committedFile) === false)
                     {
-                        violations.push(`${committedFile} matches rule ${pattern.name}`);
-                        break;
+                        rawViolations[committedFile]["content"].push(pattern.name);
+                        // NOTE: We don't break here because a file may violate > 1 rule
                     }
                 }
 
+                // Filename extension-based scanning
                 const extension = extname(committedFile).toLocaleLowerCase();
-
                 for(let k in violatingFilenameExtensions)
                 {
                     if(violatingFilenameExtensions.includes(extension) === true)
                     {
-                        violations.push(`${committedFile} matches file extension ${extension}`);
+                        rawViolations[committedFile]["extension"].push(extension);
                         break;
                     }
                 }
             }
         }
+
+// TODO -> Array.filter() 
+        // Prune any empty array items
+        let prunedViolations = [];
+        for(let violation in rawViolations)
+        {
+            const contentLength = rawViolations[violation].content.length;
+            const extensionLength = rawViolations[violation].extension.length;
+            if(contentLength > 0 || extensionLength > 0)
+            {
+                prunedViolations[violation] = rawViolations[violation];
+            }
+        }
+
+// TODO -> Array.filter()        
+        // Filter for files the user has chosen to ignore
+        for(let violation in prunedViolations)
+        {
+            let ignore = false;
+            for(let i in filesToIgnore)
+            {
+                const match = violation.match(filesToIgnore[i]);         
+                if(match)
+                {
+                    ignore = true;
+                }
+            }
+            
+            if(ignore === false)
+            {
+                filteredViolations[violation] = prunedViolations[violation];
+            }
+        }
+
     }
     catch(e)
     {
@@ -119,7 +161,7 @@ async function scanFilteredFiles(committedFiles: Set, fileContentRegexps: Array,
         }
         else
         {
-            resolve(violations);
+            resolve(filteredViolations);
         }
 
     });